requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.2
slowapi==0.1.9
raxe
requests==2.31.0

